Now compiles with lua

+++ cherry/src/events/event.cpp
@@ -21,7 +21,7 @@ void Cherry::EventListener::InitEventListenerSystem() {
                                      EventType::MouseMoveEvent,   EventType::MouseScrollEvent,
                                      EventType::KeyPressEvent,    EventType::KeyReleaseEvent };
 
-    for (int i = 0; i < types.size(); i++) {
+    for (size_t i = 0; i < types.size(); i++) {
         EventListeners[types[i]] = std::vector<Cherry::EventListener*>();
     }
 }
